Symbol fitness in SymbolLocation
